Remove non-SIMD code
