+++ .gitignore
@@ -1,6 +1,4 @@
 stb_image.h
 stb_image_write.h
 repalette
-benchmark
-benchmark_simd
 test.jpg
