+++ Makefile
@@ -11,16 +11,10 @@ WASM_FLAGS=$(WASM_CFLAGS) $(foreach flag,$(WASM_LDFLAGS),-Wl,$(flag))
 .PHONY: all
 all: repalette repalette.wasm
 
-repalette: main.c repalette_simd.c stb_image.h stb_image_write.h
-	$(CC) -o $@ main.c repalette_simd.c $(CFLAGS) $(LDLIBS) $(NATIVE_FLAGS)
+repalette: main.c repalette.c stb_image.h stb_image_write.h
+	$(CC) -o $@ main.c repalette.c $(CFLAGS) $(LDLIBS) $(NATIVE_FLAGS)
 
-benchmark: benchmark.c repalette.c stb_image.h
-	$(CC) -o $@ benchmark.c repalette.c $(CFLAGS) $(LDLIBS) $(NATIVE_FLAGS)
-
-benchmark_simd: benchmark_simd.c repalette_simd.c repalette.c stb_image.h
-	$(CC) -o $@ benchmark_simd.c repalette_simd.c repalette.c $(CFLAGS) $(LDLIBS) $(NATIVE_FLAGS)
-
-repalette.wasm: wasm_main.c repalette_simd.c
+repalette.wasm: wasm_main.c repalette.c
 	$(CC) $(WASM_FLAGS) $(CFLAGS) -o $@ $^
